+++ R/extract_stats.R
@@ -1,11 +1,14 @@
-#' @title Extracting data frames from `{ggstatsplot}` plots
+#' @title Extracting data frames or expressions from `{ggstatsplot}` plots
 #'
 #' @details
 #'
-#' This is a convenience function to extract data frames with statistical details
-#' that are used to create expressions displayed in `{ggstatsplot}` plots as
-#' subtitle, caption, etc. Note that all of this analysis is carried out by
-#' the `{statsExpressions}` [package](https://indrajeetpatil.github.io/statsExpressions/).
+#' These are convenience functions to extract data frames or expressions with
+#' statistical details that are used to create expressions displayed in
+#' `{ggstatsplot}` plots as subtitle, caption, etc. Note that all of this
+#' analysis is carried out by the `{statsExpressions}`
+#' [package](https://indrajeetpatil.github.io/statsExpressions/). And so if you
+#' are using these functions only to extract data frames, you are better off
+#' using that package.
 #'
 #' The only exception is the `ggcorrmat()` function. But, if a data frame is
 #' what you want, you shouldn't be using `ggcorrmat()` anyway. You can use
@@ -47,21 +50,23 @@
 #' }
 #' @export
 extract_stats <- function(p, ...) {
+  # styler: off
   list(
-    subtitle_data             = tryCatch(p$plot_env$subtitle_df, error = function(e) NULL),
-    caption_data              = tryCatch(p$plot_env$caption_df, error = function(e) NULL),
-    pairwise_comparisons_data = tryCatch(p$plot_env$mpc_df, error = function(e) NULL),
+    subtitle_data             = tryCatch(p$plot_env$subtitle_df,    error = function(e) NULL),
+    caption_data              = tryCatch(p$plot_env$caption_df,     error = function(e) NULL),
+    pairwise_comparisons_data = tryCatch(p$plot_env$mpc_df,         error = function(e) NULL),
     descriptive_data          = tryCatch(p$plot_env$descriptive_df, error = function(e) NULL),
-    one_sample_data           = tryCatch(p$plot_env$onesample_df, error = function(e) NULL),
-    tidy_data                 = tryCatch(p$plot_env$tidy_df, error = function(e) NULL),
-    glance_data               = tryCatch(p$plot_env$glance_df, error = function(e) NULL)
+    one_sample_data           = tryCatch(p$plot_env$onesample_df,   error = function(e) NULL),
+    tidy_data                 = tryCatch(p$plot_env$tidy_df,        error = function(e) NULL),
+    glance_data               = tryCatch(p$plot_env$glance_df,      error = function(e) NULL)
   )
+  # styler: on
 }
 
-#' @noRd
-.eval_f <- function(.f, ...) {
-  tryCatch(
-    suppressWarnings(suppressMessages(exec(.f, ...))),
-    error = function(e) NULL
-  )
-}
+#' @rdname extract_stats
+#' @export
+extract_subtitle <- function(p) extract_stats(p)$subtitle_data$expression[[1L]]
+
+#' @rdname extract_stats
+#' @export
+extract_caption <- function(p) extract_stats(p)$caption_data$expression[[1L]]