+++ tests/testthat/test-ggpiestats.R
@@ -197,9 +197,9 @@ test_that(
       data = ggplot2::msleep,
       x = conservation,
       y = vore,
-      k = 4L,
-      output = "subtitle"
-    )
+      k = 4L
+    ) %>%
+      extract_subtitle()
 
     set.seed(123)
     stats_output <- suppressWarnings(contingency_table(
@@ -283,25 +283,5 @@ test_that(
         proportion.test = FALSE
       )
     )
-  }
-)
-
-# expression output --------------------
-
-test_that(
-  desc = "expression output is as expected",
-  code = {
-    set.seed(123)
-    grouped_expr <- grouped_ggpiestats(
-      mtcars,
-      grouping.var = am,
-      x = cyl,
-      output = "subtitle"
-    )
-
-    set.seed(123)
-    base_expr <- ggpiestats(dplyr::filter(mtcars, am == "0"), cyl, output = "subtitle")
-
-    expect_equal(grouped_expr$`0`, base_expr)
   }
 )